requirements.txt
streamlit
streamlit_authenticator
pyyaml
fastapi
uvicorn
langchain
langchain_community
google-genai
pypdf
langchain-huggingface
sentence-transformers
python-dotenv
chromadb
huggingface-hub
beautifulsoup4
streamlit-cookies-controller
youtube-transcript-api
langchain-chroma
pydantic
requests